[refactor] :  ECS 렌더링 아키텍처 재설계
  ocornut/imgui master 브랜치를 extern/imgui 서브모듈로 추가, GLFW + OpenGL3
backend 와 함께 src/imgui/ STATIC 라이브러리(SJH::imgui)로 인라인 컴파일하여
project_deps 에 합류시키는 방식. 신규 apps/imguitest/ 챕터로 smoke test.

핵심 결정:
- 브랜치: master (안정), docking 미사용
- 통합: 헤더+소스 인라인 컴파일 (BuildExternLibs / lib 산출물 무관)
- 노출: project_deps 직접 합류 (모든 활성 챕터 자동 링크)
- GLSL: 프로젝트 정책 GL 4.1 Core / "#version 410 core" 강제
- 완료 기준: MSVC VS2019/VS2022 Debug/Release + macOS Ninja 모두 통과

[plan] : SP3 — ImGui 통합 구현 플랜 작성

8 태스크 분해:
1. extern/imgui 서브모듈 추가 + 안정 태그 핀
2. src/imgui/CMakeLists.txt — SJH::imgui STATIC (project_deps 합류 전)
3. SJH::imgui 를 project_deps 에 사후 합류 + 회귀 검증
4. apps/imguitest/ 챕터 — init() override GL 4.1 + ShowDemoWindow
5. macOS Release 검증
6. build-msvc.yml 매트릭스에 VS 2019 추가 + CI 4 매트릭스 검증
7. ARM64 호스트 x64 크로스 빌드 (선택)

[chore] : .gitignore 에 .worktrees/ 추가

subagent-driven 실행 시 격리된 worktree 디렉토리를 추적 대상에서 제외.

[deps] : ocornut/imgui v1.92.8 서브모듈 추가

SP3 ImGui 통합 플랜에 따라 extern/imgui 를 ocornut/imgui v1.92.8 안정
태그로 핀. 이후 src/imgui/ 에서 핵심 5 + backend 2 파일을 직접 컴파일
하는 SJH::imgui STATIC 라이브러리로 합류시킬 예정.

[dev] : add test plan

[deps] : catchorg/Catch2 v3.15.0 서브모듈 추가

테스트 프레임워크. test/ 의 코드가 v3 분할 헤더 문법
(<catch2/catch_test_macros.hpp>) 과 Catch2::Catch2WithMain 타겟을 쓰므로
v3.15.0 안정 태그로 핀. devel 브랜치 트래킹.

통합 방식: add_subdirectory(extern/Catch2) — Catch2Config / catch_discover_tests
매크로가 자동 제공되어 find_package 불필요. 추후 ENABLE_TESTING 옵션 게이트로
일반 빌드 시 컴파일 오버헤드 우회.

[deps] : Catch2 v3 통합 + test_smoke 검증 챕터

루트 CMakeLists.txt 에 ENABLE_TESTING 옵션 (기본 OFF) 추가. ON 시
add_subdirectory(extern/Catch2) 로 Catch2::Catch2WithMain 노출 +
extras/Catch.cmake 의 catch_discover_tests 매크로 사용 가능.

test_smoke/ — Catch2 통합 자체를 검증하는 최소 smoke test.
src/ 모듈에 의존하지 않으므로 통합 회귀가 즉시 드러난다.

활성화: cmake --preset ninja -DENABLE_TESTING=ON
실행:   cmake --build --preset ninja --target test_smoke
        ctest --test-dir build_ninja

[design] : SP2 — RenderContext (저수준 GL 게이트웨이) 설계 스펙 작성

ECS 렌더링 아키텍처 재설계(SP1~SP4) 의 SP2. 핵심 결정:
- Pattern Y (A6) — uniform 캐시는 Program 멤버, setter 자유 함수 유지
- CQS-clean Q2 — GetLocation pure const query, eager BuildUniformCache
- 싱글톤 RenderContext::Get() (Cocos/Godot 정통)
- Explicit Side Effects 준수 — primitive 4 개 + BeginFrame alias
- Queue 패턴 SP2 미도입 — SP3 으로 이전 (Layer A 가 ECS 도메인)
- GL 3.3 유지, DSA 미채택
- Polymorphic Program 거부 — SP3 Material 데이터 분기로 흡수

ddd rules 14 종 중 8 종에 명시 정합. SP3 가 받을 seam 정리.

[plan] : SP2 — RenderContext (저수준 GL 게이트웨이) 구현 플랜 작성

6 개 Task — Pattern Y (A6) + ddd Rules 정합 구현:

T1. src/render/ 모듈 + RenderTarget 추상 + DefaultRenderTarget
T2. RenderContext 싱글톤 + GL 게이트웨이 13 메서드 (primitive 4 + draw 2
    + BeginFrame alias + RenderTarget 위임)
T3. Program 에 uniform 캐시 멤버 + GetLocation/GetType pure const query 추가
    (additive only — Uniforms::BuildCache 호출과 과도기 공존)
T4. Uniforms 내부를 prog.GetLocation 으로 전환 + 레거시 자유함수 제거
    (sCacheRegistry / BuildCache / Forget 폐기)
T5. Program::Use() 제거 + 활성 챕터 RenderContext 마이그레이션
T6. 통합 검증 + SP3 seam 표면 확인

static_assert 6 종 (SP1 의 4 + SP2 의 2) 으로 컴파일 타임 invariant 검증.
ddd Rules 8 종에 명시 정합 (Separation/CQS/Explicit Side Effects/POLA 등).

[test] : Catch2 v3 통합 + test/ 144개 단위 테스트 활성화

ENABLE_TESTING=ON 시 add_subdirectory(test) 추가. 기존 untracked test/ 를
프로젝트 표준에 맞춰 갱신:

- glad/glad.h → GL/gl3w.h (test/*.cpp 12개 + support/gl_test_fixture.cpp)
- gladLoadGLLoader → gl3wInit (gl_test_fixture.cpp)
- GLFW_FALSE → GL_FALSE (현 GLFW 헤더에 GLFW_FALSE 없음)
- glad::glad/glfw → project_deps (sb7 의 gl3w + glfw3 INTERFACE 노출)
- fmt::fmt → spdlog::spdlog (spdlog 번들 fmt 사용, <spdlog/fmt/fmt.h>)
- cmake/Dependency.cmake 에 spdlog::spdlog ALIAS 추가

미존재 모듈 테스트 3개 비활성화 (구현 후 코멘트 해제):
- test_glfw_utils — src/input/glfw_input_utils.h 미존재
- test_scene_node — src/object/scene_node.h 미존재
- test_scene_graph — src/object/scene_graph.h 미존재

API 명명 불일치 수정:
- test_light.cpp 의 vmath::Tvec3 .x/.y/.z → [0]/[1]/[2]

test_smells linter 는 스크립트 존재 시에만 등록 (EXISTS 가드).

검증: ctest --test-dir build_ninja → 144/144 passed (100%, 10.2s)

[feat] : RenderTarget 추상 + DefaultRenderTarget + src/render/ 모듈 신설 (SP2)

RenderTarget — Bind() + GetSize() 인터페이스.
DefaultRenderTarget — glBindFramebuffer(0) + glViewport 로 화면(FBO 0) 바인딩.
SP4 에서 FrameBufferTarget 파생 추가 예정 (인터페이스 안정 유지).

[feat] : RenderContext 싱글톤 + GL 게이트웨이 13 메서드 (SP2)

Pattern Y (A6) 의 GL 호출 단일 진입점. 책임:
- mBoundProgram 추적
- glUseProgram / glBindVertexArray / glBindTexture / glClear / glEnable*
- glDrawElements / glDrawArrays
- RenderTarget 바인딩 위임

primitive 4 개(BindTarget/Clear/SetDepthTest/SetBlend) + BeginFrame 편의
wrapper 병존 — Explicit Side Effects 준수.

복사·이동 = delete 명시 (싱글톤), static_assert 2 종 검증.

[feat] : ImGui v1.53 통합 (SJH::imgui STATIC + imguitest 챕터)

본 프로젝트의 GLFW(3.0.4)와 호환되는 마지막 ImGui 태그가 v1.53 (2018-02).
v1.54+ 는 imgui_impl_glfw 가 GLFW 3.1+ cursor API(GLFWcursor 등)를
unconditional 하게 사용해 빌드 불가. extern/sb7code 는 변경 금지 제약이라
ImGui 측을 v1.53 에 핀.

v1.53 의 backend 는 GLFW + OpenGL3 결합형 (examples/opengl3_example/
imgui_impl_glfw_gl3.cpp). <GL/gl3w.h> 직접 포함 → sb7 호환. glfw3native.h
는 _WIN32 한정 include → macOS 영향 없음.

변경:
- extern/imgui v1.92.8 → v1.53 핀 변경
- src/imgui/CMakeLists.txt — 코어 3 + backend 1 파일을 SJH::imgui STATIC
  으로 인라인 컴파일. SYSTEM include + -w/W0 PRIVATE 로 소비자 -Werror 격리.
  끝에서 project_deps 에 자동 합류.
- apps/imguitest/ — smoke test 챕터. ImGui_ImplGlfwGL3_* API 사용
  (v1.53 결합형). init() 에서 GL 4.1 Core 강제.
- apps/CMakeLists.txt — 활성 챕터 _MyApp_ → imguitest

검증: macOS Debug + Release 양쪽 빌드 통과. _MyApp_ 회귀 없음.

[feat] : Program 에 uniform 캐시 멤버 + GetLocation/GetType 추가 (SP2)

Pattern Y (A6) 의 resource-attached 캐시. Additive only —
Uniforms::BuildCache 호출은 다음 Task 에서 제거 (과도기 공존).

- mUniformCache: unordered_map<string, UniformEntry>
- BuildUniformCache(): link 직후 active uniform 전체 eager 캐시
- GetLocation(name) const: pure query, 미캐시 시 -1
- GetType(name) const: pure query, 미캐시 시 0

const-callable static_assert 2 종 검증.

[refactor] : Uniforms 자유함수 내부를 prog.GetLocation 으로 전환 (SP2)

레거시 sCacheRegistry (TU-local static) 제거. 자유 함수 시그니처는 보존 —
외부 호출자(챕터, 광원 helper) 무수정. 내부 구현만 변경:

- Set* 가 prog.GetLocation 호출 (캐시 read-only)
- -1 fallback 시 glGetUniformLocation 직접 (배열 원소 등 비-active uniform)
- Uniforms::BuildCache / Forget 자유 함수 폐기 (Program 의 멤버 흡수)
- Program::Create 의 Uniforms::BuildCache 호출 제거
- ~Program 의 Uniforms::Forget 호출 제거 (mUniformCache 자동 destroy)

program_uniforms.h 의 SP1 seam 메모를 SP2 완료 상태로 정정.

[refactor] : Program::Use() 제거 + test RenderContext 로 마이그레이션 (SP2)

Pattern Y (A6) 의 마지막 단계 — glUseProgram owner 가 RenderContext.
활성 챕터(imguitest)는 prog->Use() 미사용이므로 apps/ 무변경.
test_program_uniforms 의 prog->Use() 호출을 rc.UseProgram(*prog) 로 치환.

- src/program/program.h: Use() 선언 + SP1 seam doxygen 메모 제거 +
  stale docstring (Uniforms::BuildCache/Forget 언급) SP2 완료 상태로 정리 +
  ~Program() brief 갱신
- src/program/program.cpp: Use() 정의 제거
- test/test_program_uniforms.cpp: include render_context.h 추가 + 호출 치환
- test/CMakeLists.txt: test_program_uniforms 에 SJH::render 링크 추가

시각 회귀 smoke test — imguitest 빌드 성공 확인 (GUI 미실행).

[docs] : test_program_uniforms.cpp 의 stale docstring 정정 (SP2)

Task 5 reviewer I1 후속 — 파일 헤더 doxygen 의 @c Uniforms::BuildCache
언급을 멤버 @c BuildUniformCache 로 정정. Pattern Y 흡수 후 실 동작과 정합.

[docs] : SP2 종료 cleanup — orphan UniformEntry + stale doxygen (SP2)

Final reviewer 권고 2 종 처리:
- program_uniforms.h: Uniforms::UniformEntry orphan struct 제거
  (사용처 없음 — 캐시는 Program::UniformEntry 멤버로 이전됨)
- uniform_diagnostics.h: TU-local 정적 캐시 / Uniforms::Forget 언급을
  Program::mUniformCache 멤버 / Program::~Program Invalidate 로 정정

SP3 진입 전 doc 일관성 확보.

+++ src/object/model.h
@@ -38,12 +38,12 @@ namespace SJH
         /// @brief 서브메시(RenderUnit) 개수 반환.
         int GetMeshCount() const { return (int)mRenderUnit.size(); }
         /// @brief index 번째 메시의 비소유 관찰자. Model 보다 오래 보관 금지 — owner 는 RenderUnit.
-        Mesh *GetMesh(int index) const { return mRenderUnit[index].mesh.get(); }
+        Mesh *GetMesh(int index) const { return mRenderUnit[(size_t)index].mesh.get(); }
 
         /// @brief 보유 머티리얼 개수.
         int GetMaterialCount() const { return (int)mMaterials.size(); }
         /// @brief index 번째 머티리얼의 비소유 관찰자 — 셋업 시 @c SetProgram 주입용.
-        Material *GetMaterial(int index) const { return mMaterials[index].get(); }
+        Material *GetMaterial(int index) const { return mMaterials[(size_t)index].get(); }
 
         /// @brief 모든 RenderUnit 을 순서대로 — 머티리얼 Apply 후 메시 드로우.
         void Draw() const;
